Projekt 1/requirements.txt
numpy==1.23.5
Pillow==11.1.0